requirements.txt
streamlit
tensorflow
Pillow
matplotlib
numpy
streamlit-lottie
